release: pass target to create-release-archive and create-deb scripts
This patch allowed to simplify the create-release-archive.sh and remove data
duplication between the release.yml and the script.

To simplify even more, during the release.yml job the CARGO_TARGET is
always set so that the pattern 'target/{target}/release' is consistent.

Due to the above change, the create-deb.sh need to set the --target options
as well.